extract feed store specs helper methods to extension
